requirements.txt
scikit-learn
numpy
pandas
loguru
deepspeed
transformers
datasets
accelerate
tokenizers
evaluate
wandb
tabulate
diffusers
peft
sentencepiece
scikit-build
Cython
scalene
torchmetrics
sphinx-autobuild
myst-parser
furo
munch
typer